add manage listings and user/listings relations

+++ app/Http/Controllers/ListingController.php
@@ -41,6 +41,8 @@ class ListingController extends Controller
             $formFields['logo'] = request()->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listings::create($formFields);
 
         return redirect('/')->with('success', 'Listing created!');
@@ -53,6 +55,10 @@ class ListingController extends Controller
 
     public function update(Listings $listing)
     {
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $formFields = request()->validate([
             'title' => 'required',
             'description' => 'required',
@@ -73,8 +79,17 @@ class ListingController extends Controller
 
     public function destroy(Listings $listing)
     {
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('success', 'Listing deleted!');
+    }
+
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
